adds: exception handling and unit tests

+++ src/main/java/com/challange/brokeragemanagementapi/dto/AssetDto.java
@@ -3,6 +3,8 @@ package com.challange.brokeragemanagementapi.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class AssetDto {
@@ -10,6 +12,6 @@ public class AssetDto {
         private Long id;
         private Long customerId;
         private String assetName;
-        private Double size;
-        private Double usableSize;
+        private BigDecimal size;
+        private BigDecimal usableSize;
 }
